Idler watcher example.
The callback is called on every turn of the event loop.
idle watcher to look at the watcher life cycle and see how uv_run() will
now block because a watcher is present. The idle watcher is stopped when
the count is reached and uv_run() exits since no event watchers are
active.